authserver/requirements.txt
Flask
mysql-connector-python
get_docker_secret